==> code/02_Cleaning.ipynb <==
# Data Cleaning

# Import libraries
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

import os #listdir(path)

## Part 1 - ID and Rankings

# Import data
rank_df = pd.read_csv('../data/ranked_data.csv', index_col=0)
rank_df.head()

# Check for nulls in first 3 columns
rank_df.isnull().sum()

# Drop duplicates
rank_df = rank_df.drop_duplicates()

## Part 2 - Categorical data

# Get categorical data
cat_df = pd.read_csv('../data/cat_data.csv', index_col=0)

# Check for nulls
cat_df.iloc[:,:9].isnull().sum()

# Drop excess categories
try:
    stop = cat_df.columns.get_loc('mech_1')
    cat_df = cat_df.iloc[:,:stop]
except:
    pass

# Fill nulls
cat_df = cat_df.fillna('N/A')

# Create column with list of categories
cat_df['categories'] = cat_df.iloc[:,9:14].values.tolist()

## Part 3 - Aggregating User Data

# Import dataframe
df = pd.read_csv('../data/user_ratings.csv')

def agg_ratings():
    # Create directory path
    data_path = '../data/game_ratings/'

    # Get files
    files = os.listdir(data_path)

    # Import and concat all csvs
    df = pd.DataFrame()

    # Go through csvs
    for file in files:
        new = pd.read_csv(data_path+file, index_col=0)
        df = pd.concat([df, new], ignore_index=True)
    
    return df

def agg_ratings2(data):
    # Create directory path
    data_path = '../data/game_ratings2/'

    # Get files
    files = os.listdir(data_path)

    # Go through csvs
    for file in files:
        new = pd.read_csv(data_path+file, index_col=0)
        data = pd.concat([data, new], ignore_index=True)
    
    return data

# # Fill last of df
# df = agg_ratings2(df)

def user_clean(data):
    # Drop Unnamed: 0 column
    data.drop('Unnamed: 0', axis = 1, inplace=True)

    # Drop value column
    data.drop('value', axis = 1, inplace=True)

    # Drop all nulls and duplicates
    data = data.dropna().drop_duplicates()

    # Change users to categorical numbers
    data['username'] = data['username'].astype('category').cat.codes
    
    return data

# Check for nulls
df.isnull().sum()

# Check how many games there are
len(df['id'].unique())

# Is there user overlap?
print(len(df), len(df['username'].unique()))
df['username'].value_counts()

# Get game names
names = pd.read_csv('../data/ranked_data.csv', index_col=0)
name_to_id = dict(zip(names['id'], names['name']))
rank_to_id = dict(zip(names['id'], names['collection_rank']))

# Create list relating names to id
game_names = [name_to_id[game_id] for game_id in df['id']]
game_ranks = [rank_to_id[game_id] for game_id in df['id']]
df['game'] = game_names
df['rank'] = game_ranks

## Part 4 - Final Check

# Check for any absent ids that need to be gathered
print(len(set(cat_df['id'])))
print(len(set(df['id'])))
remove = list(set(cat_df['id']) - set(df['id']))

Since the user data is the one with less ids than the categorical data, all necessary user games are accounted for.

# Export data
# rank_df.to_csv('../data/ranked_data.csv')
# cat_df.to_csv('../data/cat_data.csv')
# df.to_csv('../data/user_ratings.csv')
